# haika-api/swagger-spec/api.yaml
openapi: 3.0.1
info:
  title: Haika APIs
  description: 'This is the Rest Api Specification for Haika app'
  termsOfService: ''
  contact:
    email: taras0318@outlook.com
  version: 1.0.0
servers:
  - url: 'http://localhost:3000/api/v1'
...
paths:
...
  '/user':
    get:
      tags:
        - user
      summary: View user info
      description: View user info
      operationId: getUserInfo
      security:
        - bearerAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/GetUser'
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/InvalidToken'        
        '404':
          $ref: '#/components/responses/NotFound' 
        '500':
          $ref: '#/components/responses/ServerError' 
    put:
      tags:
        - user
      summary: User profile editor
      description: Update an existing user profile
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        description: User object that needs to be added to the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUser'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    format: int32
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/InvalidToken'        
        '404':
          $ref: '#/components/responses/NotFound'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError' 
      x-codegen-request-body-name: body
    post:
      tags:
        - user
      summary: Add a new user
      description: Add a new user
      operationId: addUser
      security:
        - bearerAuth: []
      requestBody:
        description: User object that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddUser'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: 
                    type: integer
                    format: int32
                  data:
                    type: object
                    properties:
                      token: 
                        type: string
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/InvalidToken'        
        '404':
          $ref: '#/components/responses/NotFound' 
        '405':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError' 
      x-codegen-request-body-name: body
...
components:
  responses:
    InvalidToken:
      description: Invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    AddUser:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
    GetUser:
      type: object
      properties:
        user_id:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        updated_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT